refactor

+++ WebApi/Models/Product.cs
@@ -1,9 +1,9 @@
 public class Product
 {
     public required string Sku { get; set; }
-    public required decimal PriceInCents { get; set; }
+    public required int PriceInCents { get; set; }
     public int SpecialQuantity { get; set; }
-    public decimal SpecialPrice { get; set; }
+    public int SpecialPrice { get; set; }
 
 }
